layout/base/tests/chrome/test_last_used_printer_name.html

<!DOCTYPE HTML>
<html>
<head>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body onload="run()">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

function getLastUsedPrinterName() {
  try {
      let PSSVC = Cc["@mozilla.org/gfx/printsettings-service;1"].getService(
        Ci.nsIPrintSettingsService
      );

      return PSSVC.lastUsedPrinterName;
    } catch (e) {
      Cu.reportError(e);
    }

    return null;
}

function run() {
  isnot(getLastUsedPrinterName(), null);
  SimpleTest.finish();
}

</script>
</body>
</html>
